# src/main/resources/static/api-docs.yaml
openapi: 3.1.0
info:
  title: Cinema API
  description: REST API для управления кинотеатром — фильмы, сеансы, залы, билеты, жанры и посетители.
  version: v1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
paths:
...
  /sessions:
    get:
      tags: [sessions]
      summary: Получить все сеансы (с пагинацией)
      description: Возвращает постраничный список сеансов. По умолчанию размер страницы — 3.
      operationId: getAllSessions
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Список сеансов успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseSessionDTO"
              example:
                content:
                  - id: 1
                    startTime: "2025-06-15T18:00:00"
                    movieTitle: Интерстеллар
                    hallName: Красный зал
                    price: 350.0
                metadata:
                  page: 0
                  totalPage: 3
                  size: 3
                  totalElements: 9
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sessions/{id}:
    get:
      tags: [sessions]
      summary: Получить сеанс по ID
      description: Возвращает сеанс по указанному идентификатору.
      operationId: getSessionById
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Сеанс найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDTO"
              example:
                id: 1
                startTime: "2025-06-15T18:00:00"
                movieTitle: Интерстеллар
                hallName: Красный зал
                price: 350.0
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sessions/search:
    get:
      tags: [sessions]
      summary: Поиск сеансов по названию фильма (JPQL)
      description: Возвращает постраничный список сеансов, фильтрованных по названию фильма. Используется JPQL-запрос.
      operationId: findSessionsByMovieJPQL
      parameters:
        - name: title
          in: query
          required: true
          description: Название фильма (точное совпадение)
          schema:
            type: string
            example: Интерстеллар
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Результаты поиска
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseSessionDTO"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sessions/nativesearch:
    get:
      tags: [sessions]
      summary: Поиск сеансов по названию фильма (Native SQL)
      description: Аналог /sessions/search, но использует нативный SQL-запрос вместо JPQL.
      operationId: findSessionsByMovieNative
      parameters:
        - name: title
          in: query
          required: true
          description: Название фильма (точное совпадение)
          schema:
            type: string
            example: Джокер
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Результаты поиска
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseSessionDTO"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sessions/post:
    post:
      tags: [sessions]
      summary: Создать сеанс
      description: >
        Создаёт новый сеанс. Время начала должно быть в настоящем или будущем.
        Необходимо указать ID существующих фильма и зала.
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionRequestDTO"
            example:
              startTime: "2025-07-20T19:30:00"
              movieId: 1
              hallId: 2
      responses:
        "200":
          description: Сеанс успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDTO"
              example:
                id: 10
                startTime: "2025-07-20T19:30:00"
                movieTitle: Интерстеллар
                hallName: Синий зал
                price: 500.0
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sessions/post/good:
    post:
      tags: [sessions]
      summary: Создать несколько сеансов (с транзакцией)
      description: >
        Создаёт несколько сеансов в рамках одной транзакции.
        При ошибке на 3-м элементе все изменения откатываются — демонстрация транзакционности.
      operationId: createMultipleSessionsTransactional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/SessionRequestDTO"
            example:
              - startTime: "2025-08-01T10:00:00"
                movieId: 1
                hallId: 1
              - startTime: "2025-08-01T14:00:00"
                movieId: 2
                hallId: 1
              - startTime: "2025-08-01T18:00:00"
                movieId: 1
                hallId: 2
      responses:
        "200":
          description: Все сеансы успешно созданы
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionDTO"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sessions/post/bad:
    post:
      tags: [sessions]
      summary: Создать несколько сеансов (без транзакции — демо ошибки)
      description: >
        Создаёт несколько сеансов БЕЗ транзакции. При ошибке на 3-м элементе
        первые два сеанса уже сохранены в БД и не откатятся — демонстрация
        проблемы отсутствия транзакции.
      operationId: createMultipleSessionsNoTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/SessionRequestDTO"
            example:
              - startTime: "2025-08-02T10:00:00"
                movieId: 1
                hallId: 1
              - startTime: "2025-08-02T14:00:00"
                movieId: 2
                hallId: 1
              - startTime: "2025-08-02T18:00:00"
                movieId: 1
                hallId: 2
      responses:
        "200":
          description: Часть сеансов создана (до ошибки)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionDTO"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sessions/update/{id}:
    put:
      tags: [sessions]
      summary: Обновить сеанс
      description: Обновляет данные сеанса по ID — время начала, фильм и зал.
      operationId: updateSession
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionRequestDTO"
            example:
              startTime: "2025-07-21T20:00:00"
              movieId: 2
              hallId: 1
      responses:
        "200":
          description: Сеанс успешно обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDTO"
              example:
                id: 10
                startTime: "2025-07-21T20:00:00"
                movieTitle: Джокер
                hallName: Красный зал
                price: 350.0
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sessions/delete/{id}:
    delete:
      tags: [sessions]
      summary: Удалить сеанс
      description: Удаляет сеанс по ID и очищает кэш.
      operationId: deleteSession
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Сеанс успешно удалён
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─── TICKETS ───────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      description: Уникальный идентификатор ресурса
      schema:
        type: integer
        format: int64
        example: 1
    PageParam:
      name: page
      in: query
      required: false
      description: Номер страницы (начиная с 0)
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
        example: 0
    SizeParam:
      name: size
      in: query
      required: false
      description: Количество элементов на странице (по умолчанию 3)
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 3
        example: 3
    SortParam:
      name: sort
      in: query
      required: false
      description: "Сортировка в формате: поле,направление (например: id,asc)"
      schema:
        type: array
        items:
          type: string
        example: ["id,asc"]
  responses:
    BadRequest:
      description: >
        Некорректный запрос (400). Возможные причины:
        ошибка валидации полей (`@Valid`),
        неверный формат JSON,
        неправильный тип параметра,
        нарушение бизнес-логики (например, номер места превышает вместимость зала).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validationError:
              summary: Ошибка валидации входных данных
              value:
                status: 400
                message: "Некорректный запрос: ошибка валидации входных данных"
                time: "2025-06-15T12:00:00"
            typeMismatch:
              summary: Неправильный тип параметра
              value:
                status: 400
                message: "Некорректный запрос: неправильный тип передаваемого параметра"
                time: "2025-06-15T12:00:00"
            invalidJson:
              summary: Ошибка формата JSON
              value:
                status: 400
                message: "Некорректный запрос: ошибка в формате JSON-сообщения"
                time: "2025-06-15T12:00:00"
            businessLogic:
              summary: Нарушение бизнес-логики (IllegalArgumentException)
              value:
                status: 400
                message: "Некорректный запрос: в зале Красный зал недостаточно мест"
                time: "2025-06-15T12:00:00"
    NotFound:
      description: Ресурс не найден (404). Возникает при обращении к несуществующей сущности.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 404
            message: "Фильм с ID 99 не найден"
            time: "2025-06-15T12:00:00"
    MethodNotAllowed:
      description: >
        Неправильный HTTP-метод (405).
        Возникает при обращении к эндпоинту с недопустимым методом
        (например, POST вместо GET).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 405
            message: "Неправильный тип запроса"
            time: "2025-06-15T12:00:00"
...
    InternalServerError:
      description: >
        Непредвиденная ошибка сервера (500).
        Возникает при любом необработанном исключении.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 500
            message: "Непредвиденная ошибка на сервере"
            time: "2025-06-15T12:00:00"
  schemas:
...
    SessionRequestDTO:
      type: object
      description: Тело запроса для создания или обновления сеанса
      required: [startTime, movieId, hallId]
      properties:
        startTime:
          type: string
          format: date-time
          description: "Время начала сеанса в формате ISO 8601 (настоящее или будущее время)"
          example: "2025-07-20T19:30:00"
        movieId:
          type: integer
          format: int64
          description: ID фильма, который будет показан на сеансе
          example: 1
        hallId:
          type: integer
          format: int64
          description: ID зала, в котором проводится сеанс
          example: 2
...
    SessionDTO:
      type: object
      description: Данные сеанса
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор сеанса
          example: 1
        startTime:
          type: string
          format: date-time
          description: Время начала сеанса
          example: "2025-06-15T18:00:00"
        movieTitle:
          type: string
          description: Название фильма
          example: Интерстеллар
        hallName:
          type: string
          description: Название зала
          example: Красный зал
        price:
          type: number
          format: double
          description: Цена билета (берётся из зала)
          example: 350.0
...
    Metadata:
      type: object
      description: Метаданные пагинации
      properties:
        page:
          type: integer
          format: int32
          description: Текущий номер страницы (начиная с 0)
          example: 0
        totalPage:
          type: integer
          format: int32
          description: Общее количество страниц
          example: 5
        size:
          type: integer
          format: int32
          description: Размер страницы
          example: 3
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов во всех страницах
          example: 14
...
    PageResponseSessionDTO:
      type: object
      description: Постраничный ответ со списком сеансов
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/SessionDTO"
        metadata:
          $ref: "#/components/schemas/Metadata"
...
    ErrorResponse:
      type: object
      description: Стандартный ответ при ошибке
      properties:
        status:
          type: integer
          format: int32
          description: HTTP-код ошибки
          example: 404
        message:
          type: string
          description: Человекочитаемое описание ошибки
          example: "Фильм с ID 5 не найден"
        time:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: "2025-06-15T12:00:00"